added vulcanAuto abstract class

+++ Team8375/src/main/java/org/firstinspires/ftc/team8375/Subsystems/Intake.java
@@ -33,7 +33,7 @@ public class Intake {
          deploy_right = deployRight;
     }
 
-    public void deploy(long time) {
+    public void deploy(double time) {
         if(!reset) {
             this.time.reset();
             reset = true;
